General: Add support for most unicode letters in WP backend login names
This accepts user names that contain a single script, but not mixed-script
names, such as ones that mix Latin and Cyrillic. That seemed to be closest
to the code's existing philosophy.

Since PHP 7.2 and 7.3 don't offer mb_str_split, this change leaves sites
on those versions of PHP with ASCII-only user names.
 - Existing user names continue to work unchanged even on 7.2/7.3
 - adding new non-ASCII user names requires 7.4
 - autocorrection requires 7.4 (e.g. user'@example.com -> user@example.com)